requirements.txt
Flask
boto3
pandas
scikit-learn
joblib
mlflow
watchdog
python-dotenv
requests